requirements.txt
asgiref==3.3.1
Django==3.1.5
django-leaflet==0.27.1
djangorestframework==3.12.2
djangorestframework-gis==0.17
pytz==2021.1
sqlparse==0.4.1
